+++ app/src/main/java/org/oppia/app/onboarding/OnBoardingActivity.kt
@@ -1,11 +1,11 @@
-package org.oppia.app.onBoarding
+package org.oppia.app.onboarding
 
 import android.os.Bundle
 import org.oppia.app.activity.InjectableAppCompatActivity
 import org.oppia.app.profile.ProfileActivity
 import javax.inject.Inject
 
-/** An activity that shows a temporary OnBoardinging page until the app is onborarded then navigates to [ProfileActivity]. */
+/** An activity that shows a temporary on_boarding screen until the app is on-boarded then navigates to [ProfileActivity]. */
 class OnBoardingActivity : InjectableAppCompatActivity() {
 
   @Inject lateinit var onBoardingActivityPresenter: OnBoardingActivityPresenter
@@ -16,4 +16,3 @@ class OnBoardingActivity : InjectableAppCompatActivity() {
     onBoardingActivityPresenter.handleOnCreate()
   }
 }
-
