+++ app/src/main/res/layout/on_boarding_activity.xml
@@ -6,7 +6,7 @@
   android:orientation="vertical">
 
   <Button
-    android:id="@+id/onBoard"
+    android:id="@+id/on_board"
     android:layout_width="wrap_content"
     android:layout_height="wrap_content"
     android:text="on-board" />
